Moved debug drawing and pid math to iplib.
