+++ lib/desktop/GuiWindow.cpp
@@ -12,6 +12,10 @@
 using namespace std;
 
 void glfw_key_callback(GLFWwindow *window, int key, int scancode, int action, int mode) {
+    // ignore unused parameter warning
+    ((void)scancode);
+    ((void)mode);
+
     if (key == GLFW_KEY_ESCAPE && action == GLFW_PRESS)
         glfwSetWindowShouldClose(window, GL_TRUE);
 }
