+++ lib/desktop/draw.cpp
